Core + Apps : Merged AuthHelper and AuthService in AuthManager

+++ src/battle/core/storage/FileSystemIO.class.php
@@ -43,7 +43,9 @@ class FileSystemIO{
 	}
 
 	public static function delete_file($file_path){
-		return unlink($file_path);
+		if(is_file($file_path))
+			return unlink($file_path);
+		return false;
 	}
 }
 ?>
